Module Caqti_eio

Establishing Connections for Eio without Unix

This module provides connections to the PGX database for Eio applications. For other database systems, you will need Caqti_eio_unix.

The caqti-eio library should be considered unstable for now. Eio is in active development including its API, and the Caqti interface to it may benefit from further revision before the first major Eio release lands.

type stdenv = < net : Eio.Net.t ; clock : Eio.Time.clock ; mono_clock : Eio.Time.Mono.t >

module Stream : Caqti_stream_sig.S with type 'a fiber := 'a

module Pool : sig ... end

module type CONNECTION = Caqti_connection_sig.S with type 'a fiber := 'a and type ('a, 'e) stream := ('a, 'e) Stream.t

include Caqti_connect_sig.S with type 'a fiber := 'a and type 'a with_switch := sw:Eio.Switch.t -> 'a and type 'a with_stdenv := stdenv:stdenv -> 'a and type ('a, 'e) stream := ('a, 'e) Stream.t and type ('a, 'e) pool := ('a, 'e) Pool.t and type connection = (module CONNECTION)

type connection = (module CONNECTION)

Shortcut for the connection API passed as a value.

val connect : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> sw:Eio.Switch.t -> stdenv:stdenv -> Uri.t -> (connection, [> Caqti_error.load_or_connect ]) Stdlib.result

connect uri locates and loads a driver which can handle uri, passes uri to the driver, which establish a connection and returns a first-class module implementing Caqti_connection_sig.S.

If you use preemptive threading, note that the connection must only be used from the thread where it was created.

See Database Tweaks for details about the tweaks_version parameter.

parameter tweaks_version
Declares compatibility with database tweaks introduced up to the given version of Caqti. Defaults to a conservative value. See the above reference for more info (not linked here due to odoc issue).

parameter env
If provided, this function will do a final expansion of environment variables which occurs in the query templates of the requests executed on the connection.

val with_connection : ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> stdenv:stdenv -> Uri.t -> (connection -> ('a, [> Caqti_error.load_or_connect ] as 'e) Stdlib.result) -> ('a, 'e) Stdlib.result

with_connection uri f calls connect on uri. If connect evaluates to Ok connection, with_connection passes the connection to f. Once f either evaluates to a result, or raises an exception, with_connection closes the database connection.

parameter tweaks_version
Passed to connect.

parameter env
Passed to connect.

val connect_pool : ?pool_config:Caqti_pool_config.t -> ?post_connect:(connection -> (unit, 'connect_error) Stdlib.result) -> ?env:(Caqti_driver_info.t -> string -> Caqti_query.t) -> ?tweaks_version:(int * int) -> sw:Eio.Switch.t -> stdenv:stdenv -> Uri.t -> ((connection, [> Caqti_error.connect ] as 'connect_error) Pool.t, [> Caqti_error.load ]) Stdlib.result

connect_pool uri is a pool of database connections constructed by connect uri.

Do not use pooling for connections to volatile resources like sqlite3::memory: and beware of temporary tables or other objects which may not be shared across connections to the same URI.

If you use preemptive threading, note that the connection pool must only be used from the thread where it was created. Use thread local storage to create a separate pool per thread if necessary.

parameter pool_config
Provides tuning parameters for the pool. The default is the result of a fresh call of Caqti_pool_config.default_from_env.

parameter tweaks_version
Passed to connect when creating new connections.

parameter env
Passed to connect when creating new connections.

val or_fail : ('a, [< Caqti_error.t ]) Stdlib.result -> 'a

Eliminates the error-case by raising Caqti_error.Exn.